+++ app/gui/src/project-view/components/GraphEditor/NodeWidgetTree.vue
@@ -67,16 +67,23 @@ function handleWidgetUpdates(update: WidgetUpdate) {
   selectNode()
   const edit = update.edit ?? graph.startEdit()
   if (update.portUpdate) {
-    const { value, origin } = update.portUpdate
+    const { origin } = update.portUpdate
     if (Ast.isAstId(origin)) {
-      const ast =
-        value instanceof Ast.Ast ? value
-        : value == null ? Ast.Wildcard.new(edit)
-        : undefined
-      if (ast) {
-        edit.replaceValue(origin, ast)
-      } else if (typeof value === 'string') {
-        edit.tryGet(origin)?.syncToCode(value)
+      if ('value' in update.portUpdate) {
+        const value = update.portUpdate.value
+        const ast =
+          value instanceof Ast.Ast ? value
+          : value == null ? Ast.Wildcard.new(edit)
+          : undefined
+        if (ast) {
+          edit.replaceValue(origin, ast)
+        } else if (typeof value === 'string') {
+          edit.tryGet(origin)?.syncToCode(value)
+        }
+      }
+      if ('metadata' in update.portUpdate) {
+        const { metadataKey, metadata } = update.portUpdate
+        edit.tryGet(origin)?.setWidgetMetadata(metadataKey, metadata)
       }
     } else {
       console.error(`[UPDATE ${origin}] Invalid top-level origin. Expected expression ID.`)