+++ app/ydoc-server/src/fileFormat.ts
@@ -34,15 +34,12 @@ export const nodeMetadata = z
   })
   .passthrough()
 
-export type ImportMetadata = z.infer<typeof importMetadata>
-export const importMetadata = z.object({}).passthrough()
-
 export type IdeMetadata = z.infer<typeof ideMetadata>
 export const ideMetadata = z
   .object({
     node: z.record(z.string().uuid(), nodeMetadata),
-    import: z.record(z.string(), importMetadata),
     snapshot: z.string().optional(),
+    widget: z.optional(z.record(z.string().uuid(), z.record(z.string(), z.unknown()))),
   })
   .passthrough()
   .default(() => defaultMetadata().ide)
@@ -87,6 +84,7 @@ function defaultMetadata() {
     ide: {
       node: {},
       import: {},
+      widget: {},
     },
   }
 }